价格清单行增加币种，使用时转为头币种

+++ ibas.materials.service/src/main/webapp/bsapp/goodsissue/GoodsIssueEditApp.ts
@@ -141,8 +141,12 @@ namespace materials {
                     actions: [ibas.emMessageAction.YES, ibas.emMessageAction.NO],
                     onCompleted(action: ibas.emMessageAction): void {
                         if (action === ibas.emMessageAction.YES) {
-                            that.editData.delete();
-                            that.saveData();
+                            if (that.editData.referenced === ibas.emYesNo.YES) {
+                                that.proceeding(ibas.emMessageType.WARNING, ibas.i18n.prop("shell_data_referenced", that.editData.toString()));
+                            } else {
+                                that.editData.delete();
+                                that.saveData();
+                            }
                         }
                     }
                 });
@@ -249,7 +253,11 @@ namespace materials {
                             this.editData.goodsIssueLines.remove(item);
                         } else {
                             // 非新建标记删除
-                            item.delete();
+                            if (item.referenced === ibas.emYesNo.YES) {
+                                this.proceeding(ibas.emMessageType.WARNING, ibas.i18n.prop("shell_data_referenced", item.toString()));
+                            } else {
+                                item.delete();
+                            }
                         }
                     }
                 }
@@ -261,7 +269,7 @@ namespace materials {
             private chooseGoodsIssueLineMaterial(caller: bo.GoodsIssueLine, type?: string, filterConditions?: ibas.ICondition[]): void {
                 let that: this = this;
                 let condition: ibas.ICondition;
-                let conditions: ibas.IList<ibas.ICondition> = app.conditions.material.create();
+                let conditions: ibas.IList<ibas.ICondition> = app.conditions.material.create(this.editData.documentDate);
                 // 添加输入条件
                 if (filterConditions instanceof Array && filterConditions.length > 0) {
                     if (conditions.length > 1) {
