+++ ibas.materials.service/src/main/webapp/bsapp/goodsreceipt/GoodsReceiptEditApp.ts
@@ -164,8 +164,12 @@ namespace materials {
                     actions: [ibas.emMessageAction.YES, ibas.emMessageAction.NO],
                     onCompleted(action: ibas.emMessageAction): void {
                         if (action === ibas.emMessageAction.YES) {
-                            that.editData.delete();
-                            that.saveData();
+                            if (that.editData.referenced === ibas.emYesNo.YES) {
+                                that.proceeding(ibas.emMessageType.WARNING, ibas.i18n.prop("shell_data_referenced", that.editData.toString()));
+                            } else {
+                                that.editData.delete();
+                                that.saveData();
+                            }
                         }
                     }
                 });
@@ -272,7 +276,11 @@ namespace materials {
                             this.editData.goodsReceiptLines.remove(item);
                         } else {
                             // 非新建标记删除
-                            item.delete();
+                            if (item.referenced === ibas.emYesNo.YES) {
+                                this.proceeding(ibas.emMessageType.WARNING, ibas.i18n.prop("shell_data_referenced", item.toString()));
+                            } else {
+                                item.delete();
+                            }
                         }
                     }
                 }
@@ -284,7 +292,7 @@ namespace materials {
             private chooseGoodsReceiptLineMaterial(caller: bo.GoodsReceiptLine, filterConditions?: ibas.ICondition[]): void {
                 let that: this = this;
                 let condition: ibas.ICondition;
-                let conditions: ibas.IList<ibas.ICondition> = app.conditions.material.create();
+                let conditions: ibas.IList<ibas.ICondition> = app.conditions.material.create(this.editData.documentDate);
                 // 添加输入条件
                 if (filterConditions instanceof Array && filterConditions.length > 0) {
                     if (conditions.length > 1) {
@@ -367,12 +375,12 @@ namespace materials {
                 });
             }
             /** 更改行价格 */
-            private changePurchaseOrderItemPrice(priceList: number | ibas.Criteria, items?: bo.GoodsReceiptLine[]): void {
+            private changePurchaseOrderItemPrice(priceList: number | ibas.ICriteria, items?: bo.GoodsReceiptLine[]): void {
                 if (ibas.objects.isNull(items)) {
                     items = this.editData.goodsReceiptLines.filterDeleted();
                 }
                 if (typeof priceList === "number" && ibas.numbers.valueOf(priceList) !== 0) {
-                    let criteria: ibas.Criteria = new ibas.Criteria();
+                    let criteria: ibas.ICriteria = materials.app.conditions.materialprice.create(this.editData.documentDate);
                     let condition: ibas.ICondition = criteria.conditions.create();
                     condition.alias = materials.app.conditions.materialprice.CONDITION_ALIAS_PRICELIST;
                     condition.value = priceList.toString();
